+++ refactoring/Pokemon/PokedexActivity_PokemonView_Location.cpp
@@ -218,7 +218,7 @@ bool PokedexActivity_PokemonView_Location::renderListItems(SDL_Surface* surf_dis
     //Render list item background/surface
     listEntrySurface = SDL_CreateRGBSurfaceWithFormat(
         0,
-        surf_display->w * 0.9,
+        static_cast<int>(surf_display->w * 0.9),
         itemHeight,
         DEPTH,
         SDL_PIXELFORMAT_RGBA32
@@ -230,9 +230,9 @@ bool PokedexActivity_PokemonView_Location::renderListItems(SDL_Surface* surf_dis
 
     int spacing = 18; 
     SDL_Rect listEntryRect;
-    listEntryRect.x = surf_display->w - (surf_display->w * 0.45) - 8;
+    listEntryRect.x = static_cast<int>(surf_display->w - (surf_display->w * 0.45) - 8);
     listEntryRect.y = 65 + (i * (itemHeight + spacing));
-    listEntryRect.w = surf_display->w * 0.45;
+    listEntryRect.w = static_cast<int>(surf_display->w * 0.45);
     listEntryRect.h = itemHeight;
     if (offset + i == selectedIndex) {
         SDL_SetRenderDrawColor(renderer, 0, 0, 0, 255); // Black
@@ -264,8 +264,8 @@ bool PokedexActivity_PokemonView_Location::renderListItems(SDL_Surface* surf_dis
     SDL_Rect moveNameRect;
     moveNameRect.x = listEntryRect.x + (listEntryRect.w / 2) - (routeNameSurface->w / 2);
     moveNameRect.y = listEntryRect.y;
-    moveNameRect.w = routeNameSurface->w * 0.8;
-    moveNameRect.h = routeNameSurface->h * 0.8;
+    moveNameRect.w = static_cast<int>(routeNameSurface->w * 0.8);
+    moveNameRect.h = static_cast<int>(routeNameSurface->h * 0.8);
     PokeSurface::onDrawScaled(surf_display, routeNameSurface, &moveNameRect);
     SDL_FreeSurface(routeNameSurface);
